Addressing feedback in PR
