+++ Assets/Scripts/Prototype/Parts/PartCollection.cs
@@ -13,7 +13,7 @@ namespace Kosmos.Prototype.Parts
             public PartBase _childPart;
             public PartSocket _childSocket;
         }
-        public HashSet<PartBase> AllParts => _allParts;
+        public IReadOnlyCollection<PartBase> AllParts => _allParts;
 
         private Dictionary<PartBase, List<PartLink>> _partLinks = new();
         private HashSet<PartBase> _allParts = new();
