+++ Assets/Scripts/Prototype/VAB/PartsToEcsManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5a2615ecacd6f6b4f82c7d607c7bb888
